move audio method to utterance
